ruby 1.8.7 support
Untested, you may experience bugs or worse.

This branch may be updated from time to time,
but will not contain the latest features.

Some features are not able to be backported to 1.8.7.

Pressure those who are forcing you to use 1.8.7,
be it your boss, Apple, a third party vendor,
or even your grandmother to upgrade to 1.9.3.